refactoring silence and about to add targetdown silence
