+++ helm/robusta/values.yaml
@@ -136,7 +136,6 @@ builtinPlaybooks:
   - on_prometheus_alert:
       alert_name: TargetDown
   actions:
-  - default_enricher: {}
   - target_down_dns_enricher: {}
   stop: true
