+++ playbooks/robusta_playbooks/silence.py
@@ -1,35 +1,29 @@
 import json
 from datetime import datetime
-from enum import Enum
-from typing import Dict, List, Optional
+from typing import List
 from uuid import UUID
 
 import requests
 from pydantic import BaseModel
-
 from robusta.api import (
     ActionException,
-    ActionParams,
-    AlertManagerDiscovery,
+    AddSilenceParams,
+    BaseSilenceParams,
+    DeleteSilenceParams,
     ErrorCodes,
     ExecutionBaseEvent,
     MarkdownBlock,
-    ServiceDiscovery,
+    SilenceMatcher,
+    SilenceOperation,
     TableBlock,
     action,
+    add_silence_to_alert_manager,
+    silence_gen_headers,
+    silence_get_alertmanager_url,
+    silence_get_url_path,
 )
 
 # ref to api https://github.com/prometheus/alertmanager/blob/main/api/v2/openapi.yaml
-
-
-class Matcher(BaseModel):
-    # https://github.com/prometheus/alertmanager/blob/main/api/v2/models/matcher.go
-    isEqual: bool = (
-        True  # support old version matchers with omitted isEqual https://github.com/prometheus/alertmanager/pull/2603
-    )
-    isRegex: bool
-    name: str
-    value: str
 
 
 class SilenceStatus(BaseModel):
@@ -45,7 +39,7 @@ class Silence(BaseModel):
     createdBy: str
     startsAt: datetime
     endsAt: datetime
-    matchers: List[Matcher]
+    matchers: List[SilenceMatcher]
 
     def to_list(self) -> List[str]:
         return [
@@ -59,52 +53,16 @@ class Silence(BaseModel):
         ]
 
 
-class BaseSilenceParams(ActionParams):
-    """
-    :var alertmanager_url: Alternative Alert Manager url to send requests.
-    """
-
-    alertmanager_flavor: str = None  # type: ignore
-    alertmanager_url: Optional[str]
-    grafana_api_key: str = None  # type: ignore
-
-
-class DeleteSilenceParams(BaseSilenceParams):
-    """
-    :var id: uuid of the silence.
-    """
-
-    id: str
-
-
-class AddSilenceParams(BaseSilenceParams):
-    """
-    :var id: uuid of the silence. use for update, empty on create.
-    :var comment: text comment of the silence.
-    :var createdBy: author of the silence.
-    :var startsAt: date.
-    :var endsAt: date.
-    :var matchers: List of matchers to filter the silence effect.
-    """
-
-    id: Optional[str]
-    comment: str
-    createdBy: str
-    startsAt: datetime
-    endsAt: datetime
-    matchers: List[Matcher]
-
-
 @action
 def get_silences(event: ExecutionBaseEvent, params: BaseSilenceParams):
-    alertmanager_url = _get_alertmanager_url(params)
+    alertmanager_url = silence_get_alertmanager_url(params)
     if not alertmanager_url:
         raise ActionException(ErrorCodes.ALERT_MANAGER_DISCOVERY_FAILED)
 
     try:
         response = requests.get(
-            f"{alertmanager_url}{_get_url_path(SilenceOperation.LIST, params)}",
-            headers=_gen_headers(params),
+            f"{alertmanager_url}{silence_get_url_path(SilenceOperation.LIST, params)}",
+            headers=silence_gen_headers(params),
         )
     except Exception as e:
         raise ActionException(ErrorCodes.ALERT_MANAGER_REQUEST_FAILED) from e
@@ -127,23 +85,7 @@ def get_silences(event: ExecutionBaseEvent, params: BaseSilenceParams):
 
 @action
 def add_silence(event: ExecutionBaseEvent, params: AddSilenceParams):
-    alertmanager_url = _get_alertmanager_url(params)
-    if not alertmanager_url:
-        raise ActionException(ErrorCodes.ALERT_MANAGER_DISCOVERY_FAILED)
-
-    try:
-        res = requests.post(
-            f"{alertmanager_url}{_get_url_path(SilenceOperation.CREATE, params)}",
-            data=params.json(exclude_defaults=True),  # support old versions.
-            headers=_gen_headers(params),
-        )
-    except Exception as e:
-        raise ActionException(ErrorCodes.ALERT_MANAGER_REQUEST_FAILED) from e
-
-    if not res.ok:
-        raise ActionException(ErrorCodes.ADD_SILENCE_FAILED, msg=f"Add silence failed: {res.text}")
-
-    silence_id = res.json().get("silenceID") or res.json().get("id")  # on grafana alertmanager the 'id' is returned
+    silence_id = add_silence_to_alert_manager(params)
     if not silence_id:
         raise ActionException(ErrorCodes.ADD_SILENCE_FAILED)
 
@@ -160,16 +102,16 @@ def add_silence(event: ExecutionBaseEvent, params: AddSilenceParams):
 
 @action
 def delete_silence(event: ExecutionBaseEvent, params: DeleteSilenceParams):
-    alertmanager_url = _get_alertmanager_url(params)
+    alertmanager_url = silence_get_alertmanager_url(params)
     if not alertmanager_url:
         raise ActionException(ErrorCodes.ALERT_MANAGER_DISCOVERY_FAILED)
 
     try:
-        alertmanager_url = _get_alertmanager_url(params)
+        alertmanager_url = silence_get_alertmanager_url(params)
 
         requests.delete(
-            f"{alertmanager_url}{_get_url_path(SilenceOperation.DELETE, params)}/{params.id}",
-            headers=_gen_headers(params),
+            f"{alertmanager_url}{silence_get_url_path(SilenceOperation.DELETE, params)}/{params.id}",
+            headers=silence_gen_headers(params),
         )
     except Exception as e:
         raise ActionException(ErrorCodes.ALERT_MANAGER_REQUEST_FAILED) from e
@@ -185,34 +127,3 @@ def delete_silence(event: ExecutionBaseEvent, params: DeleteSilenceParams):
     )
 
 
-SilenceOperation = Enum("SilenceOperation", "CREATE DELETE LIST")
-
-
-def _gen_headers(params: BaseSilenceParams) -> Dict:
-    headers = {"Content-type": "application/json"}
-    if params.grafana_api_key:
-        headers.update({"Authorization": "Bearer {0}".format(params.grafana_api_key)})
-    return headers
-
-
-def _get_url_path(operation: SilenceOperation, params: BaseSilenceParams) -> str:
-    prefix = ""
-    if "grafana" == params.alertmanager_flavor:
-        prefix = "/api/alertmanager/grafana"
-
-    if operation == SilenceOperation.DELETE:
-        return f"{prefix}/api/v2/silence"
-    else:
-        return f"{prefix}/api/v2/silences"
-
-
-def _get_alertmanager_url(params: BaseSilenceParams) -> str:
-    if params.alertmanager_url:
-        return params.alertmanager_url
-
-    if "grafana" == params.alertmanager_flavor:
-        return ServiceDiscovery.find_url(
-            selectors=["app.kubernetes.io/name=grafana"], error_msg="Failed to find grafana url"
-        )
-
-    return AlertManagerDiscovery.find_alert_manager_url()
