+++ src/robusta/api/__init__.py
@@ -1,10 +1,13 @@
 from robusta.core.external_apis.prometheus.models import PrometheusQueryResult
 from robusta.core.model.base_params import (
     ActionParams,
+    AddSilenceParams,
     AlertResourceGraphEnricherParams,
+    BaseSilenceParams,
     BashParams,
     ChartValuesFormat,
     CustomGraphEnricherParams,
+    DeleteSilenceParams,
     EventEnricherParams,
     FindingKeyParams,
     GrafanaAnnotationParams,
@@ -19,6 +22,7 @@ from robusta.core.model.base_params import (
     ResourceChartItemType,
     ResourceChartResourceType,
     ResourceGraphEnricherParams,
+    SilenceMatcher,
     TimedPrometheusParams,
     VideoEnricherParams,
 )
@@ -97,6 +101,13 @@ from robusta.core.playbooks.prometheus_enrichment_utils import (
     create_resource_enrichment,
     get_node_internal_ip,
     run_prometheus_query,
+)
+from robusta.core.playbooks.silence_utils import (
+    SilenceOperation,
+    add_silence_to_alert_manager,
+    silence_gen_headers,
+    silence_get_alertmanager_url,
+    silence_get_url_path,
 )
 from robusta.core.playbooks.trigger import (
     BaseTrigger,
